Split templates into separate CSS/JS and HTML files
