+++ internal/dataentry/handlers_test.go
@@ -52,7 +52,7 @@ func newHandlerTestApp(t *testing.T) *App {
 	}
 
 	styleMap, styledTypes := buildStyleMap(cfg, meta)
-	tmpl, err := template.New("").Funcs(templateFuncs(styleMap, styledTypes)).Parse(allTemplates)
+	tmpl, err := template.New("").Funcs(templateFuncs(styleMap, styledTypes)).Parse(allTemplates())
 	if err != nil {
 		t.Fatalf("parsing templates: %v", err)
 	}
