+++ internal/dataentry/templates/base.html
@@ -1,0 +1,113 @@
+{{- define "head" -}}
+<meta charset="utf-8">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<link rel="icon" type="image/svg+xml" href="/static/favicon.svg">
+<script>
+(function(){var t=localStorage.getItem('theme');if(t){document.documentElement.setAttribute('data-theme',t)}else if(matchMedia('(prefers-color-scheme:dark)').matches){document.documentElement.setAttribute('data-theme','dark')}})();
+</script>
+<script src="/static/htmx.min.js"></script>
+<link rel="stylesheet" href="/static/easymde.min.css">
+<script src="/static/easymde.min.js"></script>
+<link rel="stylesheet" href="/static/slimselect.css">
+<script src="/static/slimselect.min.js"></script>
+<script src="/static/mermaid.min.js"></script>
+<link rel="stylesheet" href="/static/app.css">
+<script src="/static/app.js"></script>
+{{- end -}}
+
+{{- define "nav-item" -}}
+{{ if .Dashboard }}
+    <a href="/dashboard"{{ if eq "_dashboard" .ActiveList }} class="active"{{ end }}
+       hx-get="/dashboard" hx-target="#content" hx-push-url="true">
+      {{ .Label }}
+    </a>
+{{ else if .Graph }}
+    <a href="/graph"{{ if eq "_graph" .ActiveList }} class="active"{{ end }}>
+      {{ .Label }}
+    </a>
+{{ else if .Kanban }}
+    <a href="/kanban/{{ .Kanban }}"{{ if eq (printf "_kanban_%s" .Kanban) .ActiveList }} class="active"{{ end }}
+       hx-get="/kanban/{{ .Kanban }}" hx-target="#content" hx-push-url="true">
+      {{ .Label }}
+    </a>
+{{ else }}
+    <a href="/list/{{ .List }}"{{ if eq .List .ActiveList }} class="active"{{ end }}
+       data-entity-type="{{ .EntityType }}"
+       hx-get="/list/{{ .List }}" hx-target="#content" hx-push-url="true">
+      {{ .Label }}<span class="nav-count">{{ .Count }}</span>
+    </a>
+{{ end }}
+{{- end -}}
+
+{{- define "conflict-bar" -}}
+{{ if and .ConflictCount (gt .ConflictCount 0) }}
+<div class="conflict-banner" id="conflict-banner">
+  <div>
+    <span class="conflict-banner-icon">⚠️</span>
+    <strong>{{ .ConflictCount }} file{{ if gt .ConflictCount 1 }}s{{ end }} with merge conflicts</strong> —
+    <a href="/conflicts" hx-get="/conflicts" hx-target="#content" hx-push-url="true">Resolve now</a>
+  </div>
+  <button class="conflict-banner-close" onclick="this.parentElement.remove();document.body.classList.remove('has-conflict-banner');" title="Dismiss">×</button>
+</div>
+<script>document.body.classList.add('has-conflict-banner');</script>
+{{ end }}
+{{- end -}}
+
+{{- define "sidebar" -}}
+<aside class="sidebar">
+  <div class="sidebar-header">
+    <h1>{{ .App.Name }}</h1>
+    {{ if .App.Description }}<p>{{ .App.Description }}</p>{{ end }}
+  </div>
+  <nav>
+    <a href="/search"{{ if eq $.ActiveList "_search" }} class="active"{{ end }}
+       hx-get="/search" hx-target="#content" hx-push-url="true">
+      &#128269; Search <kbd style="margin-left:auto;">/</kbd>
+    </a>
+    <a href="/analyze"{{ if eq $.ActiveList "_analyze" }} class="active"{{ end }}
+       hx-get="/analyze" hx-target="#content" hx-push-url="true"
+       style="border-bottom:1px solid rgba(255,255,255,0.1);margin-bottom:4px;">
+      &#9888; Analysis
+    </a>
+    {{ range .Navigation }}
+    {{ if .Group }}
+    <div class="nav-group">
+      <button class="nav-group-toggle" onclick="toggleNavGroup(this)" data-group="{{ .Group.Group }}">
+        <span class="nav-group-chevron{{ if .Group.Collapsed }} collapsed{{ end }}">&#9656;</span>
+        {{ .Group.Group }}
+      </button>
+      <div class="nav-group-items{{ if .Group.Collapsed }} hidden{{ end }}">
+        {{ range .Group.Items }}
+        {{ template "nav-item" (map "Dashboard" .Dashboard "Graph" .Graph "Kanban" .Kanban "Label" .Label "List" .List "EntityType" .EntityType "Count" .Count "ActiveList" $.ActiveList) }}
+        {{ end }}
+      </div>
+    </div>
+    {{ else }}
+    {{ template "nav-item" (map "Dashboard" .Item.Dashboard "Graph" .Item.Graph "Label" .Item.Label "List" .Item.List "EntityType" .Item.EntityType "Count" .Item.Count "ActiveList" $.ActiveList) }}
+    {{ end }}
+    {{ end }}
+  </nav>
+</aside>
+<div class="status-bar">
+  <div class="status-bar-left">
+    <a href="/settings"{{ if eq $.ActiveList "_settings" }} class="active"{{ end }}
+       hx-get="/settings" hx-target="#content" hx-push-url="true">&#9881; Settings</a>
+  </div>
+  <div class="status-bar-right">
+    {{ if .GitEnabled }}
+    <button class="git-status" id="git-status-btn" onclick="openSyncModal()" title="Sync with remote">
+      <span class="git-branch" id="git-branch">{{ .GitBranch }}</span>
+      <span class="git-indicator" id="git-indicator">
+        <span class="git-dot"></span>
+        <span id="git-status-text">Synced</span>
+      </span>
+    </button>
+    {{ end }}
+    <button onclick="_toggleShortcuts()"><kbd>?</kbd> Shortcuts</button>
+  </div>
+</div>
+<button class="theme-toggle" onclick="toggleTheme()" title="Toggle dark mode">
+  <span class="icon-sun">&#9788;</span>
+  <span class="icon-moon">&#9790;</span>
+</button>
+{{- end -}}
