+++ internal/dataentry/templates/kanban.html
@@ -1,0 +1,106 @@
+{{- define "kanban-page" -}}
+<!DOCTYPE html>
+<html lang="en">
+<head>
+<title>{{ .App.Name }} - {{ .Kanban.Title }}</title>
+{{ template "head" . }}
+</head>
+<body>
+{{ template "sidebar" . }}
+<main class="main main-wide" id="content">
+{{ template "kanban-content" . }}
+</main>
+<div id="command-toast-container"></div>
+</body>
+</html>
+{{- end -}}
+
+{{- define "kanban-content" -}}
+<script>document.getElementById('content').classList.add('main-wide');</script>
+<div class="page-header">
+  <div>
+    <h2>{{ .Kanban.Title }}</h2>
+  </div>
+  <div style="display:flex;gap:8px;align-items:center;">
+    <span style="font-size:13px;color:var(--text-muted);">{{ .TotalCount }} items</span>
+    {{ if .CreateForm }}
+    <a id="kanban-new-btn" href="/form/{{ .CreateForm }}" class="btn btn-primary btn-sm"
+       hx-get="/form/{{ .CreateForm }}" hx-target="#content" hx-push-url="true">+ New <kbd>N</kbd></a>
+    {{ end }}
+  </div>
+</div>
+
+{{ if .FilterControls }}
+<div class="filter-bar-sentinel"></div>
+<div class="filter-bar">
+  {{ range .FilterControls }}
+  <div>
+    <label>{{ .Label }}</label>
+    <select name="filter_{{ .Property }}" onchange="applyKanbanFilter(this, '{{ $.KanbanID }}')">
+      <option value="">All</option>
+      {{ $cur := .Current }}
+      {{ range .Values }}<option value="{{ . }}"{{ if eq . $cur }} selected{{ end }}>{{ . }}</option>{{ end }}
+    </select>
+  </div>
+  {{ end }}
+</div>
+{{ end }}
+
+{{ if .HasSwimlanes }}
+<div class="kanban-board with-swimlanes" data-kanban-id="{{ .KanbanID }}"
+     style="grid-template-columns: auto repeat({{ len .Columns }}, minmax(240px, 1fr));">
+  <div class="kanban-swimlane-header">
+    <div class="kanban-corner"></div>
+    {{ range .Columns }}
+    <div class="kanban-col-header">{{ .Label }}</div>
+    {{ end }}
+  </div>
+  {{ range $lane := .Swimlanes }}
+  <div class="kanban-swimlane" data-swimlane="{{ $lane.Value }}">
+    <div class="kanban-swimlane-label">{{ $lane.Label }}</div>
+    {{ range $col := $.Columns }}
+    <div class="kanban-cell" data-column="{{ $col.Value }}" data-swimlane="{{ $lane.Value }}">
+      {{ $cards := index (index $.Cells $col.Value) $lane.Value }}
+      {{ range $cards }}
+      {{ template "kanban-card" (map "Card" . "EditForm" $.EditForm "KanbanID" $.KanbanID "FilterParams" $.FilterParams) }}
+      {{ end }}
+    </div>
+    {{ end }}
+  </div>
+  {{ end }}
+</div>
+{{ else }}
+<div class="kanban-board" data-kanban-id="{{ .KanbanID }}">
+  {{ range $col := .Columns }}
+  <div class="kanban-column" data-column="{{ $col.Value }}">
+    <div class="kanban-column-header">
+      <span>{{ $col.Label }}</span>
+      {{ $cards := index (index $.Cells $col.Value) "" }}
+      <span class="kanban-count">{{ len $cards }}</span>
+    </div>
+    <div class="kanban-cards">
+      {{ range $cards }}
+      {{ template "kanban-card" (map "Card" . "EditForm" $.EditForm "KanbanID" $.KanbanID "FilterParams" $.FilterParams) }}
+      {{ end }}
+    </div>
+  </div>
+  {{ end }}
+</div>
+{{ end }}
+{{- end -}}
+
+{{- define "kanban-card" -}}
+<div class="kanban-card" draggable="true" data-entity-id="{{ .Card.ID }}"
+     data-accent="{{ accentColor .Card.AccentType .Card.AccentValue }}"
+     hx-get="/form/{{ .EditForm }}/{{ .Card.ID }}?from=/kanban/{{ .KanbanID }}{{ .FilterParams }}"
+     hx-target="#content" hx-push-url="true">
+  <div class="kanban-card-title">{{ .Card.Title }}</div>
+  {{ if .Card.Fields }}
+  <div class="kanban-card-fields">
+    {{ range .Card.Fields }}
+    <span class="badge {{ badgeClass .PropType .Value }}">{{ .Value }}</span>
+    {{ end }}
+  </div>
+  {{ end }}
+</div>
+{{- end -}}
