+++ internal/dataentry/templates/list.html
@@ -1,0 +1,124 @@
+{{- define "page" -}}
+<!DOCTYPE html>
+<html lang="en">
+<head>
+<title>{{ .App.Name }} - {{ .List.Title }}</title>
+{{ template "head" . }}
+</head>
+<body>
+{{ template "sidebar" . }}
+{{ template "conflict-bar" . }}
+<main class="main" id="content">
+{{ template "list-content" . }}
+</main>
+<div id="command-toast-container"></div>
+</body>
+</html>
+{{- end -}}
+
+{{- define "list-content" -}}
+<div class="page-header">
+  <div>
+    <h2>{{ .List.Title }}</h2>
+    {{ if .List.Description }}<p>{{ .List.Description }}</p>{{ end }}
+  </div>
+  <div style="display:flex;gap:8px;align-items:center;">
+    <span style="font-size:13px;color:var(--text-muted);">{{ .TotalCount }} items</span>
+    {{ if .List.CreateForm }}
+    <a href="/form/{{ .List.CreateForm }}" class="btn btn-primary btn-sm"
+       hx-get="/form/{{ .List.CreateForm }}" hx-target="#content" hx-push-url="true">+ New <kbd>N</kbd></a>
+    {{ end }}
+    {{ if .Commands }}{{ if gt (len .Commands) 2 }}
+    <details class="add-dropdown">
+      <summary class="btn btn-secondary btn-sm">Commands &#9662;</summary>
+      <div class="add-dropdown-menu">
+        {{ range .Commands }}<a href="#" onclick="event.preventDefault();runCommand('{{ .ID }}', {list_id:'{{ $.ListID }}'})" {{ if .Confirm }}data-confirm="{{ .Confirm }}"{{ end }}{{ if boolTrue .AutoOpen }} data-auto-open="true"{{ end }}>{{ .Label }}</a>
+        {{ end }}
+      </div>
+    </details>
+    {{ else }}{{ range .Commands }}
+    <button class="btn btn-secondary btn-sm" onclick="runCommand('{{ .ID }}', {list_id:'{{ $.ListID }}'})" {{ if .Confirm }}data-confirm="{{ .Confirm }}"{{ end }}{{ if boolTrue .AutoOpen }} data-auto-open="true"{{ end }}>{{ .Label }}</button>
+    {{ end }}{{ end }}{{ end }}
+  </div>
+</div>
+
+{{ if .FilterControls }}
+<div class="filter-bar">
+  {{ range .FilterControls }}
+  <div>
+    <label>{{ .Label }}</label><br>
+    {{ if or (eq .Widget "select") (eq .Widget "multi-select") }}
+    <select name="filter_{{ .Property }}"
+            hx-get="/list/{{ $.ListID }}" hx-target="#content" hx-push-url="true"
+            hx-include=".filter-bar select, .filter-bar input">
+      <option value="">All</option>
+      {{ $current := .Current }}
+      {{ range .Values }}<option value="{{ . }}"{{ if eq . $current }} selected{{ end }}>{{ . }}</option>{{ end }}
+    </select>
+    {{ else }}
+    <input type="text" placeholder="Search..." name="filter_{{ .Property }}"
+           hx-get="/list/{{ $.ListID }}" hx-target="#content" hx-push-url="true"
+           hx-trigger="keyup changed delay:300ms"
+           hx-include=".filter-bar select, .filter-bar input">
+    {{ end }}
+  </div>
+  {{ end }}
+</div>
+{{ end }}
+
+{{ if .List.Filters }}
+<div class="info-bar">
+  {{ range .List.Filters }}
+  <span class="info-chip">{{ .Property }} {{ .Operator }} {{ .Value }}</span>
+  {{ end }}
+</div>
+{{ end }}
+
+<div class="card">
+  <div style="overflow-x:auto;">
+    <table>
+      <thead>
+        <tr>
+          {{ range .Columns }}<th{{ if .Sortable }} class="sortable" hx-get="{{ .SortURL }}" hx-target="#content" hx-push-url="true" hx-include=".filter-bar select, .filter-bar input"{{ end }}>{{ if .Label }}{{ .Label }}{{ else }}{{ .Property }}{{ end }}{{ if .IsSorted }}<span class="sort-indicator">{{ if eq .SortDir "desc" }}&#9660;{{ else }}&#9650;{{ end }}</span>{{ end }}</th>{{ end }}
+          <th></th>
+        </tr>
+      </thead>
+      <tbody>
+        {{ range .Rows }}
+        <tr{{ if $.EditForm }} data-edit-href="/form/{{ $.EditForm }}/{{ .EntityID }}?return_to=/list/{{ $.ListID }}"{{ end }}>
+          {{ $dlp := $.DetailLinkPrefix }}
+          {{ range .Cells }}
+          <td>
+            {{ if .Link }}<a href="{{ $dlp }}{{ .EntityID }}?from={{ $.ListID }}&scope=list:{{ $.ListID }}{{ $.ScopeParams }}" class="cell-link"
+               hx-get="{{ $dlp }}{{ .EntityID }}?from={{ $.ListID }}&scope=list:{{ $.ListID }}{{ $.ScopeParams }}" hx-target="#content" hx-push-url="true">{{ .Value }}</a>
+            {{ else if isBadgeType .PropType }}<span class="badge {{ badgeClass .PropType .Value }}">{{ .Value }}</span>
+            {{ else }}{{ if .Value }}{{ .Value }}{{ else }}&mdash;{{ end }}{{ end }}
+          </td>
+          {{ end }}
+          <td style="width:1%;white-space:nowrap;"><a href="#" class="delete-icon" title="Delete"
+              onclick="event.preventDefault();confirmDelete('{{ .EntityID }}','/list/{{ $.ListID }}')"><svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><polyline points="3 6 5 6 21 6"/><path d="M19 6v14a2 2 0 01-2 2H7a2 2 0 01-2-2V6m3 0V4a2 2 0 012-2h4a2 2 0 012 2v2"/></svg></a></td>
+        </tr>
+        {{ end }}
+        {{ if not .Rows }}
+        <tr><td colspan="{{ add (len .Columns) 1 }}" style="text-align:center;padding:32px;color:var(--text-muted);">No items found</td></tr>
+        {{ end }}
+      </tbody>
+    </table>
+  </div>
+  <div class="pagination">
+    <span>{{ .TotalCount }} items{{ if .HasPagination }} &middot; Page {{ .Page }} of {{ .TotalPages }}{{ end }}</span>
+    {{ if .HasPagination }}
+    <div style="display:flex;gap:6px;">
+      {{ if .PrevPageURL }}<a href="{{ .PrevPageURL }}" class="btn btn-secondary btn-sm"
+         hx-get="{{ .PrevPageURL }}" hx-target="#content" hx-push-url="true"
+         hx-include=".filter-bar select, .filter-bar input">&larr; Prev <kbd>H</kbd></a>
+      {{ else }}<span class="btn btn-secondary btn-sm" style="opacity:0.4;pointer-events:none;">&larr; Prev <kbd>H</kbd></span>{{ end }}
+      {{ if .NextPageURL }}<a href="{{ .NextPageURL }}" class="btn btn-secondary btn-sm"
+         hx-get="{{ .NextPageURL }}" hx-target="#content" hx-push-url="true"
+         hx-include=".filter-bar select, .filter-bar input">Next &rarr; <kbd>L</kbd></a>
+      {{ else }}<span class="btn btn-secondary btn-sm" style="opacity:0.4;pointer-events:none;">Next &rarr; <kbd>L</kbd></span>{{ end }}
+    </div>
+    {{ end }}
+  </div>
+</div>
+{{- end -}}
